chatgpt-translate/requirements.txt
openai==0.27.2
pyyaml
